+++ packages/admin-stories/src/admin-react-select/FinalFormReactSelect.tsx
@@ -1,6 +1,6 @@
-import { Field, FinalFormInput } from "@comet/admin";
+import { Field, FinalFormInput, FormPaper } from "@comet/admin";
 import { FinalFormReactSelectStaticOptions } from "@comet/admin-react-select";
-import { Button, ListItem, Typography } from "@material-ui/core";
+import { Box, Button } from "@material-ui/core";
 import { storiesOf } from "@storybook/react";
 import * as React from "react";
 import { Form } from "react-final-form";
@@ -12,12 +12,12 @@ function Story() {
         { value: "vanilla", label: "Vanilla" },
     ];
     return (
-        <div style={{ width: "300px" }}>
-            <ListItem>
+        <FormPaper variant="outlined" style={{ width: 300 }}>
+            <Box marginBottom={4}>
                 <Button component={"button"} disableTouchRipple>
-                    <Typography variant="button">blah</Typography>
+                    blah
                 </Button>
-            </ListItem>
+            </Box>
 
             <Form
                 onSubmit={(values) => {
@@ -37,7 +37,7 @@ function Story() {
                     </form>
                 )}
             />
-        </div>
+        </FormPaper>
     );
 }