+++ packages/admin-stories/src/admin-rte/FieldAllOptions.tsx
@@ -1,6 +1,6 @@
-import { Field } from "@comet/admin";
+import { Field, FormPaper } from "@comet/admin";
 import { createFinalFormRte } from "@comet/admin-rte";
-import { Button, Typography } from "@material-ui/core";
+import { Box, Button, Typography } from "@material-ui/core";
 import { storiesOf } from "@storybook/react";
 import * as React from "react";
 import { Form } from "react-final-form";
@@ -16,30 +16,34 @@ function Story() {
     const [submitedValue, setSubmittedValue] = React.useState<{ rteContent: any }>({ rteContent: defaultContent });
 
     return (
-        <div>
-            <div style={{ maxWidth: "800px" }}>
-                <Form
-                    initialValues={{
-                        rteContent: defaultContent,
-                    }}
-                    onSubmit={(values) => {
-                        //
-                        setSubmittedValue({ rteContent: values.rteContent });
-                    }}
-                    render={({ handleSubmit }) => (
-                        <form onSubmit={handleSubmit}>
-                            <Field name="rteContent" label="Rich Text" component={RteField} />
-                            <Button color="primary" variant="contained" type="submit" component={"button"} disableTouchRipple>
-                                <Typography variant="button">Submit</Typography>
-                            </Button>
-                        </form>
-                    )}
-                />
-            </div>
-            <Typography variant="h5" color="primary">
-                Readonly Component:
-            </Typography>
-            <RteReadOnly content={submitedValue.rteContent} />
+        <div style={{ maxWidth: 800 }}>
+            <Box marginBottom={4}>
+                <FormPaper variant="outlined">
+                    <Form
+                        initialValues={{
+                            rteContent: defaultContent,
+                        }}
+                        onSubmit={(values) => {
+                            //
+                            setSubmittedValue({ rteContent: values.rteContent });
+                        }}
+                        render={({ handleSubmit }) => (
+                            <form onSubmit={handleSubmit}>
+                                <Field name="rteContent" label="Rich Text" component={RteField} />
+                                <Button color="primary" variant="contained" type="submit" component={"button"} disableTouchRipple>
+                                    <Typography variant="button">Submit</Typography>
+                                </Button>
+                            </form>
+                        )}
+                    />
+                </FormPaper>
+            </Box>
+            <FormPaper variant="outlined">
+                <Typography variant="h5" color="primary">
+                    Readonly Component:
+                </Typography>
+                <RteReadOnly content={submitedValue.rteContent} />
+            </FormPaper>
         </div>
     );
 }