+++ packages/admin-stories/src/admin/error-handling/RouteWithErrorBoundary.tsx
@@ -1,5 +1,5 @@
 import { MasterLayout, Menu, MenuItemRouterLink, RouteWithErrorBoundary } from "@comet/admin";
-import { Typography } from "@material-ui/core";
+import { Box, Paper, Typography } from "@material-ui/core";
 import { Alert } from "@material-ui/lab";
 import { storiesOf } from "@storybook/react";
 import * as React from "react";
@@ -8,14 +8,17 @@ import StoryRouter from "storybook-react-router";
 
 const ViewWithNoError: React.FunctionComponent = () => {
     return (
-        <div>
-            <Typography variant={"h5"}>View with No Error</Typography>
-
-            <Alert severity={"info"}>
-                <Typography>Use the RouteWithErrorBoundary component to create a boundary for app routes.</Typography>
-                <Typography>Try to click on the second route (Error Route) to display a route which throws an error.</Typography>
-            </Alert>
-        </div>
+        <Paper variant="outlined">
+            <Box padding={4}>
+                <Typography variant={"h5"} gutterBottom>
+                    View with No Error
+                </Typography>
+                <Alert severity={"info"}>
+                    <Typography>Use the RouteWithErrorBoundary component to create a boundary for app routes.</Typography>
+                    <Typography>Try to click on the second route (Error Route) to display a route which throws an error.</Typography>
+                </Alert>
+            </Box>
+        </Paper>
     );
 };