+++ packages/admin-stories/src/admin/mui/Buttons.tsx
@@ -39,7 +39,7 @@ export const Story: React.FC = () => {
     return (
         <Grid container spacing={4}>
             <Grid item xs={12}>
-                <Paper>
+                <Paper variant="outlined">
                     <Box padding={4}>
                         <Typography variant={"h2"}>Contained buttons</Typography>
                     </Box>
@@ -88,7 +88,7 @@ export const Story: React.FC = () => {
                 </Paper>
             </Grid>
             <Grid item xs={12}>
-                <Paper>
+                <Paper variant="outlined">
                     <Box padding={4}>
                         <Typography variant={"h2"}>Button groups</Typography>
                     </Box>
@@ -131,7 +131,7 @@ export const Story: React.FC = () => {
                 </Paper>
             </Grid>
             <Grid item xs={12}>
-                <Paper>
+                <Paper variant="outlined">
                     <Box padding={4}>
                         <Typography variant={"h2"}>Text buttons</Typography>
                     </Box>
@@ -180,7 +180,7 @@ export const Story: React.FC = () => {
                 </Paper>
             </Grid>
             <Grid item xs={6}>
-                <Paper>
+                <Paper variant="outlined">
                     <Box padding={4}>
                         <Typography variant={"h2"}>Special button usage</Typography>
                     </Box>
@@ -241,7 +241,7 @@ export const Story: React.FC = () => {
                 </Paper>
             </Grid>
             <Grid item xs={6}>
-                <Paper>
+                <Paper variant="outlined">
                     <Box padding={4}>
                         <Typography variant={"h2"}>Icon buttons</Typography>
                         <Table>