+++ packages/admin-stories/src/admin/table/TableDndOrder.tsx
@@ -3,6 +3,8 @@ import { Button, Typography } from "@material-ui/core";
 import { storiesOf } from "@storybook/react";
 import * as React from "react";
 
+import { dndProviderDecorator } from "../../dnd.decorator";
+
 interface IRow {
     id: string; // TODO add support for number in TableLocalChanges
     order: number;
@@ -56,4 +58,6 @@ function Story() {
     );
 }
 
-storiesOf("@comet/admin/table", module).add("DnD Order", () => <Story />);
+storiesOf("@comet/admin/table", module)
+    .addDecorator(dndProviderDecorator())
+    .add("DnD Order", () => <Story />);