+++ packages/admin/src/form/Field.tsx
@@ -1,5 +1,6 @@
+import { FieldValidator } from "final-form";
 import * as React from "react";
-import { Field as FinalFormField, FieldRenderProps } from "react-final-form";
+import { Field as FinalFormField, FieldRenderProps, FormSpy, useForm } from "react-final-form";
 
 import { FieldContainer, FieldContainerThemeProps } from "./FieldContainer";
 
@@ -8,32 +9,38 @@ const requiredValidator = (value: any) => (value ? undefined : "Pflichtfeld");
 const composeValidators = (...validators: Array<(value: any, allValues: object) => any>) => (value: any, allValues: object) =>
     validators.reduce((error, validator) => error || validator(value, allValues), undefined);
 
-interface IVividFieldProps<FieldValue = any, T extends HTMLElement = HTMLElement> {
+interface Props<FieldValue = any, T extends HTMLElement = HTMLElement> {
     name: string;
     label?: React.ReactNode;
     component?: React.ComponentType<any> | string;
     children?: (props: FieldRenderProps<FieldValue, T>) => React.ReactNode;
     required?: boolean;
     disabled?: boolean;
-    validate?: (value: any, allValues: object) => any;
+    validate?: FieldValidator<FieldValue>;
+    validateWarning?: FieldValidator<FieldValue>;
     variant?: FieldContainerThemeProps["variant"];
     [otherProp: string]: any;
 }
 
-export class Field<FieldValue = any, T extends HTMLElement = HTMLElement> extends React.Component<IVividFieldProps<FieldValue, T>> {
-    public render() {
-        const { children, component, name, label, required, validate, ...rest } = this.props;
-        const composedValidate = required ? (validate ? composeValidators(requiredValidator, validate) : requiredValidator) : validate;
-        return (
-            <FinalFormField<FieldValue, FieldRenderProps<FieldValue, T>, T> name={name} validate={composedValidate} {...rest}>
-                {this.renderField.bind(this)}
-            </FinalFormField>
-        );
-    }
+export function Field<FieldValue = any, FieldElement extends HTMLElement = HTMLElement>({
+    children,
+    component,
+    name,
+    label,
+    required,
+    validate,
+    validateWarning,
+    ...otherProps
+}: Props<FieldValue, FieldElement>): React.ReactElement {
+    const { disabled, variant, fullWidth } = otherProps;
 
-    private renderField({ input, meta, fieldContainerProps, ...rest }: FieldRenderProps<FieldValue, T>) {
-        const { children, component, name, label, required, disabled, variant, fullWidth } = this.props;
+    const { mutators } = useForm();
+    const setFieldData = mutators.setFieldData as ((...args: any[]) => any) | undefined;
+    const currentWarningValidationRound = React.useRef(0);
 
+    const validateError = required ? (validate ? composeValidators(requiredValidator, validate) : requiredValidator) : validate;
+
+    function renderField({ input, meta, fieldContainerProps, ...rest }: FieldRenderProps<FieldValue, FieldElement> & { warning?: string }) {
         function render() {
             if (component) {
                 return React.createElement(component, { ...rest, input, meta });
@@ -49,7 +56,8 @@ export class Field<FieldValue = any, T extends HTMLElement = HTMLElement> extend
                 label={label}
                 required={required}
                 disabled={disabled}
-                error={(meta.error || meta.submitError) && meta.touched && (meta.error || meta.submitError)}
+                error={meta.touched && (meta.error || meta.submitError)}
+                warning={meta.touched && meta.data?.warning}
                 variant={variant}
                 fullWidth={fullWidth}
             >
@@ -57,4 +65,41 @@ export class Field<FieldValue = any, T extends HTMLElement = HTMLElement> extend
             </FieldContainer>
         );
     }
+
+    return (
+        <>
+            <FinalFormField<FieldValue, FieldRenderProps<FieldValue, FieldElement>, FieldElement>
+                name={name}
+                validate={validateError}
+                {...otherProps}
+            >
+                {renderField}
+            </FinalFormField>
+            {validateWarning && (
+                <FormSpy
+                    subscription={{ values: true }}
+                    onChange={async ({ values }) => {
+                        if (!setFieldData) {
+                            console.warn(
+                                `Can't perform validateWarning, as the setFieldData mutator is missing. Did you forget to add the mutator to the form?`,
+                            );
+                            return;
+                        }
+
+                        currentWarningValidationRound.current++;
+                        const validationRound = currentWarningValidationRound.current;
+
+                        const warning = await Promise.resolve(validateWarning(values[name], values));
+
+                        if (currentWarningValidationRound.current > validationRound) {
+                            // Another validation has been started, skip this one
+                            return;
+                        }
+
+                        setFieldData(name, { warning });
+                    }}
+                />
+            )}
+        </>
+    );
 }