+++ packages/admin/src/helpers/PrettyBytes.tsx
@@ -1,0 +1,32 @@
+import { NumberFormatOptions } from "@formatjs/ecma402-abstract";
+import * as React from "react";
+import { CustomFormatConfig, FormattedNumber } from "react-intl";
+
+type AvailableUnits = "byte" | "kilobyte" | "megabyte" | "gigabyte" | "terabyte" | "petabyte";
+const availableUnits: AvailableUnits[] = ["byte", "kilobyte", "megabyte", "gigabyte", "terabyte", "petabyte"];
+
+interface PrettyBytesProps extends Omit<NumberFormatOptions & CustomFormatConfig, "style"> {
+    value: number;
+    unit?: AvailableUnits;
+}
+
+export const PrettyBytes = ({
+    value: bytes,
+    unit: customUnit,
+    maximumFractionDigits: customMaximumFractionDigits,
+    ...props
+}: PrettyBytesProps): React.ReactElement => {
+    let exponent;
+    if (customUnit) {
+        exponent = availableUnits.findIndex((unit) => unit === customUnit);
+    } else {
+        // log to base 1024
+        // => receive exponent of equation `1024 ** exponent = bytes`
+        // => exponent determines unit
+        exponent = Math.min(Math.floor(Math.log(Math.abs(bytes)) / Math.log(1024)), availableUnits.length - 1);
+    }
+    const unit = availableUnits[exponent];
+    const size = bytes / 1024 ** exponent;
+
+    return <FormattedNumber value={size} style="unit" unit={unit} maximumFractionDigits={customMaximumFractionDigits ?? 2} {...props} />;
+};