+++ packages/admin/src/index.ts
@@ -127,6 +127,9 @@ export { ErrorDialogProvider } from "./error/errordialog/ErrorDialogProvider";
 export { useErrorDialog } from "./error/errordialog/useErrorDialog";
 export { ErrorScope, LocalErrorScopeApolloContext, errorScopeForOperationContext } from "./error/errordialog/ErrorScope";
 export { ErrorDialogApolloLinkOptions, createErrorDialogApolloLink } from "./error/errordialog/createErrorDialogApolloLink";
+export { FilterBar } from "./table/filterbar/FilterBar";
+export { FilterBarPopoverFilter } from "./table/filterbar/FilterBarPopoverFilter";
+export { FilterBarMoreFilters } from "./table/filterbar/FilterBarMoreFilters";
 export { SplitButton, SplitButtonProps } from "./common/buttons/split/SplitButton";
 export { CometAdminSplitButtonThemeProps } from "./common/buttons/split/SplitButton.styles";
 export { SaveButton, SaveButtonProps } from "./common/buttons/save/SaveButton";
@@ -147,3 +150,4 @@ export {
     CometAdminCometAdminFinalFormSaveCancelButtonsLegacyClassKeys,
     CometAdminCometAdminFinalFormSaveCancelButtonsLegacyThemeProps,
 } from "./FinalFormSaveCancelButtonsLegacy.styles";
+export { PrettyBytes } from "./helpers/PrettyBytes";