+++ packages/admin/src/table/filterbar/index.ts
@@ -1,0 +1,3 @@
+export * from "./FilterBar";
+export * from "./FilterBarPopoverFilter";
+export * from "./FilterBarMoreFilters";