+++ packages/admin/src/themeAugmentation.ts
@@ -9,6 +9,8 @@ import { CometAdminMenuCollapsibleItemClassKeys, MenuCollapsibleItemThemeProps }
 import { CometAdminMenuItemClassKeys } from "./mui/menu/Item";
 import { MenuThemeProps } from "./mui/menu/Menu";
 import { CometAdminMenuClassKeys } from "./mui/menu/Menu.styles";
+import { CometAdminFilterBarClassKeys } from "./table/filterbar/FilterBar";
+import { CometAdminFilterBarPopoverFilterClassKeys, FilterBarPopoverFilterThemeProps } from "./table/filterbar/FilterBarPopoverFilter";
 
 declare module "@material-ui/core/styles/overrides" {
     interface ComponentNameToClassKey {
@@ -22,6 +24,8 @@ declare module "@material-ui/core/styles/overrides" {
         CometAdminFormPaper: CometAdminFormPaperKeys;
         CometAdminFormSection: CometAdminFormSectionKeys;
         CometAdminFinalFormRangeInput: CometAdminFinalFormRangeInputClassKeys;
+        CometAdminFilterBar: CometAdminFilterBarClassKeys;
+        CometAdminFilterBarPopoverFilter: CometAdminFilterBarPopoverFilterClassKeys;
     }
 }
 
@@ -32,5 +36,6 @@ declare module "@material-ui/core/styles/props" {
         CometAdminMenu: MenuThemeProps;
         CometAdminMenuCollapsibleItem: MenuCollapsibleItemThemeProps;
         CometAdminErrorBoundary: ErrorBoundaryThemeProps;
+        CometAdminFilterBarPopoverFilter: FilterBarPopoverFilterThemeProps;
     }
 }